@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\AnyOf\AnyOf.csproj" />
-  </ItemGroup>
-
   <!--<ItemGroup>
-    <PackageReference Include="AnyOf" Version="0.0.20" />
+    <ProjectReference Include="..\..\src\AnyOf\AnyOf.csproj" />
   </ItemGroup>-->
+
+  <ItemGroup>
+    <PackageReference Include="AnyOf" Version="0.4.0" />
+  </ItemGroup>
 
 </Project>
